requirements.txt
astroid==2.4.2
blessings==1.7
bpython==0.19
certifi==2020.6.20
chardet==3.0.4
curtsies==0.3.4
greenlet==0.4.16
idna==2.10
isort==4.3.21
JPype1==1.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.0
peewee==3.13.3
Pillow==7.2.0
Pygments==2.6.1
pylint==2.5.3
pypng==0.0.20
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
